requirements.txt
django==3.2.10
djangorestframework==3.13.0
django-allauth==0.13.0
django-rest-auth==0.9.5
phonenumbers==8.12.42
psycopg2-binary==2.9.3
pillow==8.4.0
djangorestframework-simplejwt==5.0.0
autopep8==1.6.0
autoflake==1.4
isort==5.10.1
drf-yasg==1.20.0
PyJWT==2.3.0
